+++ Collaboration3_iBanProject/Login/LoginView.swift
@@ -9,45 +9,49 @@ import SwiftUI
 import Firebase
 
 struct LoginView: View {
+    // MARK: - Properties
     @State private var email: String = ""
     @State private var password: String = ""
+    @State private var isLoginEnabled: Bool = false
     
+    // MARK: - Body
     var body: some View {
-        VStack {
+        content
+    }
+    
+    // MARK: - Content
+    private var content: some View {
+        VStack(spacing: 20) {
             Spacer()
-            Image("logoWithTitle")
-                .resizable()
-                .aspectRatio(contentMode: .fit)
-                .frame(width: 150, height: 150)
-                .padding(.bottom, 60)
-            
+            LogoImage()
             EmailTextField(email: $email)
-        
             PasswordSecureField(password: $password)
-            
-            HStack {
-                Text("Don't have an account?")
-                    .foregroundColor(.white)
-                
-                Button(action: {
-                    // Action for signup button
-                }) {
-                    Text("Sign Up")
-                        .foregroundColor(.customAccentColor)
-                }
-            }
-            .padding(.top, 20)
-            
+            accountCheckView
             Spacer()
-            LoginButton(onTap: login)
-            
+            AuthActionButtonView(actionText: "Log In", isEnabled: isLoginEnabled, onTap: login)
             Spacer()
         }
         .background(Color.customBackgroundColor)
         .edgesIgnoringSafeArea(.all)
     }
     
-    func login() {
+    private var accountCheckView: some View {
+        HStack {
+            Text("Don't have an account?")
+                .foregroundColor(.white)
+            
+            Button(action: {
+                // Action for signup button
+            }) {
+                Text("Sign Up")
+                    .foregroundColor(.customAccentColor)
+            }
+        }
+        .padding(.top, 20)
+    }
+    
+    // MARK: - Methods
+    private func login() {
         Auth.auth().signIn(withEmail: email, password: password) { result, error in
             if error != nil {
                 print(error!.localizedDescription)
@@ -55,63 +59,6 @@ struct LoginView: View {
         }
     }
 }
-
-struct EmailTextField: View {
-    @Binding var email: String
-    
-    var body: some View {
-        HStack {
-            Text("Email")
-                .foregroundColor(.white)
-            Spacer()
-        }
-        .padding(.horizontal)
-        TextField("Enter your email address", text: $email)
-            .padding()
-            .background(Color.customTextFieldColor)
-            .tint(.white)
-            .cornerRadius(8.0)
-            .padding(.horizontal)
-    }
-}
-
-struct PasswordSecureField: View {
-    @Binding var password: String
-    
-    var body: some View {
-        HStack {
-            Text("Password")
-                .foregroundColor(.white)
-            Spacer()
-        }
-        .padding(.horizontal)
-        
-        SecureField("Enter your password", text: $password)
-            .padding()
-            .background(Color.customTextFieldColor)
-            .cornerRadius(8.0)
-            .padding(.horizontal)
-    }
-}
-
-struct LoginButton: View {
-    var onTap: () -> Void
-
-    var body: some View {
-        Button(action: {
-            onTap()
-        }) {
-            Text("Log In")
-                .foregroundColor(.white)
-                .padding()
-                .frame(maxWidth: .infinity)
-                .background(Color.customAccentColor)
-                .cornerRadius(8.0)
-                .padding(.horizontal)
-        }
-    }
-}
-
 
 #Preview {
     LoginView()
